place measurements and actions back into one module, subclassing Device
