+++ bellini/devices.py
@@ -16,13 +16,26 @@ from bellini.units import VOLUME_UNIT
 # BASE CLASS
 # =============================================================================
 
-class ActionableDevice(abc.ABC):
+class Device(abc.ABC):
+    """ Base class for devices (objects that don't change over a procedure) """
+
+class ActionableDevice(Device):
     """ Base class for object that can manipulate experiment objects """
 
     @abc.abstractmethod
     def apply(self, *args, **kwargs):
         """
         Manipulate the provided experimental objects and return the new objects
+        """
+        raise NotImplementedError
+
+class MeasurementDevice(Device):
+    """ Base function for measurement instruments """
+
+    @abc.abstractmethod
+    def readout(self, *args, **kwargs):
+        """
+        Return measurement(s) of the given objects (generally with some error)
         """
         raise NotImplementedError
 
@@ -64,3 +77,26 @@ class LiquidTransfer(ActionableDevice):
         new_sink = sink.receive_aliquot(aliquot)
 
         return new_source, new_sink
+
+class Measurer(MeasurementDevice):
+    """ Measure a property of one container with Gaussian error """
+    def __init__(self, name, var):
+        """
+        TODO: allow variance to be drawn from a prior, update name for multiple
+        measurements
+        """
+
+        if isinstance(var, Quantity):
+            self.var = var
+        else:
+            raise ValueError("var must be a Quantity")
+
+        self.name = name
+        self.measure_count = 0
+
+    def readout(self, container, value, key=None):
+        measurement = Normal(container.observe(value, key=key), self.var)
+        measurement.name = f"{self.name}_{container}_{value}_{self.measure_count}"
+        self.measure_count += 1
+        measurement.observed = True
+        return measurement
